server/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
APScheduler==3.6.3
coverage==5.3
Django==3.0.5
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.11.1
djongo==1.3.3
factory-boy==3.1.0
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psutil==5.7.3
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pymongo==3.11.0
pytest==6.1.1
pytest-cov==2.10.1
pytest-django==4.0.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
uuid==1.30
wrapt==1.12.1
numpy==1.19.3
pandas==1.1.4
colorlog==4.7.2